# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dealna backend API",
        "title": "Dealna API",
        "contact": {
            "name": "Amro Deek"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/purchases/items/{itemId}/request": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new purchase request from the authenticated user to buy a specific item",
                "tags": [
                    "Purchases"
                ],
                "summary": "Send a purchase request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PurchaseRequest"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/purchases/items/{itemId}/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of all purchase requests made for a specific item",
                "tags": [
                    "Purchases"
                ],
                "summary": "List purchase requests for an item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.PurchaseRequest"
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/purchases/items/{itemId}/requests/{requestId}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows the seller to accept a specific purchase request",
                "tags": [
                    "Purchases"
                ],
                "summary": "Accept a purchase request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/purchases/items/{itemId}/requests/{requestId}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows the buyer to cancel their own purchase request",
                "tags": [
                    "Purchases"
                ],
                "summary": "Cancel a purchase request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/purchases/items/{itemId}/requests/{requestId}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows the seller to reject a specific purchase request",
                "tags": [
                    "Purchases"
                ],
                "summary": "Reject a purchase request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item ID",
                        "name": "itemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/purchases/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of all purchase requests made by the authenticated user",
                "tags": [
                    "Purchases"
                ],
                "summary": "Get my purchase requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.PurchaseRequest"
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.PurchaseRequest": {
            "type": "object",
            "properties": {
                "buyerID": {
                    "type": "string"
                },
                "buyerName": {
                    "type": "string"
                },
                "buyerPic": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "itemID": {
                    "type": "string"
                },
                "requestID": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.PurchaseRequestStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "domain.PurchaseRequestStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "ACCEPTED",
                "REJECTED",
                "FROZEN",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "PurchaseRequestPending",
                "PurchaseRequestAccepted",
                "PurchaseRequestRejected",
                "PurchaseRequestFrozen",
                "PurchaseRequestCancelled"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}